Improve error logging with stack traces
Prior to, errors encountered only logged the type and message, but
didn't include a stack trace. This can be very frustrating when
debugging, particularly when just starting to use hubot.

Stack traces are our friends :)

And since we're in the neighborhood, also update to use a
consistent variable for the error object.